+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- The `huggingface_hub` safetensors API has changed, so we did not fetch the number of
+  model parameters correctly - this has now been fixed.
+
 ## [v16.14.0] - 2026-02-13
 
 ### Added